+++ docs/src/components/SubCategory.svelte
@@ -83,11 +83,13 @@
   }
 
   .body {
+    overflow: hidden;
     margin-bottom: 0;
     max-height: 0;
     overflow-y: hidden;
     transition:
       max-height 250ms cubic-bezier(0.4, 0, 0.2, 1),
+      height 250ms cubic-bezier(0.4, 0, 0.2, 1),
       margin-bottom 250ms cubic-bezier(0.4, 0, 0.2, 1);
   }