+++ docs/src/utils/clickOutside.ts
@@ -1,8 +1,11 @@
 /** Dispatch event on click outside of node */
-export function clickOutside(node: Node) {
-  const handleClick = (event: MouseEvent) => {
-    if (node && !node.contains(event.target) && !event.defaultPrevented) {
-      node.dispatchEvent(new CustomEvent("click_outside", node));
+export function clickOutside(node: HTMLElement) {
+  const handleClick = (event: Event) => {
+    if (
+      event.target && node && !node.contains(event.target as Node) &&
+      !event.defaultPrevented
+    ) { // @ts-ignore: node type invalid
+      node.dispatchEvent(new CustomEvent<HTMLElement>("click_outside", node));
     }
   };