+++ docs/src/utils/parsed.ts
@@ -19,6 +19,8 @@ export interface FunctionPage {
   examples?: string[];
   realm?: string;
   internal?: boolean;
+  deprecated?: boolean;
+  stub?: boolean;
   item: "function";
 }
 
@@ -34,6 +36,8 @@ export interface TablePage {
   fields?: FieldInfo[];
   realm?: string;
   internal?: boolean;
+  deprecated?: boolean;
+  stub?: boolean;
   item: "table";
 }
 
@@ -46,6 +50,9 @@ export interface Category {
   subcategories: {
     [key: string]: ValidSubcategory;
   };
+  internal?: boolean;
+  deprecated?: boolean;
+  stub?: boolean;
   item: "category";
 }