+++ docs/src/views/Function.svelte
@@ -3,17 +3,20 @@
 
   import GetTypes from "../components/GetTypes.svelte";
   import function_name from "../utils/funName";
-  import type { FunctionParameters, FunctionReturns } from "../utils/parsed";
+  import type {
+    FunctionPage,
+    FunctionParameters,
+    FunctionReturns,
+  } from "../utils/parsed";
   import Page from "./Page.svelte";
+  import FunctionSignature from "./fragments/FunctionSignature.svelte";
 
-  export let item: any;
+  export let item: FunctionPage;
   export let category: string;
 
   let title: string = "";
 
   let description = item.description || "";
-  let func_name: string = "";
-  let isMethod: boolean;
 
   const examples: string[] = item.examples ?? [];
   const parameters: FunctionParameters[] = item.parameters ?? [];
@@ -21,6 +24,9 @@
 
   const internalMessage =
     "**This is used internally - although you're able to use it you probably shouldn't.**";
+  const deprecatedMessage =
+    "**We advice agains't using this. It may be changed or removed in a future update.**";
+  const stubMessage = "**This article is a stub.**";
 
   $: {
     if (item.internal) {
@@ -28,7 +34,21 @@
         ? `${description}\n\n${internalMessage}`
         : internalMessage;
     }
+
+    if (item.deprecated) {
+      description = description
+        ? `${description}\n\n${deprecatedMessage}`
+        : deprecatedMessage;
+    }
+
+    if (item.stub) {
+      description = description
+        ? `${description}\n\n${stubMessage}`
+        : stubMessage;
+    }
   }
+  let func_name: string = "";
+  let isMethod: boolean;
 
   $: {
     [func_name, isMethod] = function_name(item.name);
@@ -40,34 +60,7 @@
 
 <Page {title}>
   <div class="section-container">
-    <div class="function-signature">
-      <span class={`realm ${item.realm || ""}`}></span>
-
-      <span class="funct">
-        {#each returns as ret}
-          <GetTypes types={ret.type} list={true} />
-        {/each}
-        {#if returns.length > 0}
-          <span>{" "}</span>
-        {/if}
-        {#if isMethod}
-          <GetTypes types={category} />
-          <span>{":"}</span>
-        {/if}
-        <span>{func_name}</span>
-
-        <span>(</span>
-
-        {#each parameters as param, i}
-          <GetTypes types={param.type} list={true} /><span
-            >{" " + param.name}</span
-          >
-          {parameters.length > 1 && i < parameters.length - 1 ? ", " : ""}
-        {/each}
-
-        <span>)</span>
-      </span>
-    </div>
+    <FunctionSignature {item} {category} />
   </div>
 
   {#if description}
@@ -158,50 +151,5 @@
   .parameter-box > p.title {
     display: inline-flex;
     margin: 0 0 0 3.8rem;
-  }
-
-  .function-signature {
-    display: inline-flex;
-    padding: 1.6rem;
-    color: var(--codeBlocks-color);
-    font-family: var(--code);
-    font-size: 1.6rem;
-    font-weight: 400;
-    line-height: 1.8rem;
-    margin: 0.8rem 0;
-    background: var(--codeBlocks-background);
-  }
-
-  .realm {
-    flex: 0 0 1.8rem;
-    height: 1.8rem;
-    width: 1.8rem;
-    display: inline-block;
-    align-self: center;
-    vertical-align: top;
-    margin-right: 0.8rem;
-    border-radius: 0.4rem;
-  }
-
-  .realm:global(.server) {
-    background: var(--realms-server);
-  }
-
-  .realm:global(.client) {
-    background: var(--realms-client);
-  }
-
-  .realm.shared {
-    background: linear-gradient(
-      45deg,
-      var(--realms-client) 50%,
-      var(--realms-server) 50.001%
-    );
-  }
-
-  .funct {
-    display: inline-flex;
-    white-space: pre;
-    flex-flow: wrap;
   }
 </style>