+++ parser/src/classes/Project/CategoryProject.ts
@@ -36,6 +36,8 @@ export default class CategoryProject {
     this.add_tag(new AliasTag("constructor", 0, ["@category globals"]));
     this.add_tag(new Tag("field", 3, false));
     this.add_tag(new BooleanTag("internal"));
+    this.add_tag(new BooleanTag("deprecated"));
+    this.add_tag(new BooleanTag("stub"));
     this.add_tag(
       new AliasTag("hook", 2, ([subcategory, name]) => [
         "@category hooks",