+++ parser/src/classes/Project/TablePage.ts
@@ -12,6 +12,8 @@ export default class TablePage {
   readonly fields?: FieldInfo[];
   readonly realm?: string;
   readonly internal?: boolean;
+  readonly deprecated?: boolean;
+  readonly stub?: boolean;
 
   constructor(
     public readonly name: string,
@@ -28,6 +30,10 @@ export default class TablePage {
     this.fields = fields.length > 0 ? fields : undefined;
     this.realm = get_unique(block, "realm");
     this.internal = get_unique(block, "internal") === "true" ? true : undefined;
+    this.deprecated = get_unique(block, "deprecated") === "true"
+      ? true
+      : undefined;
+    this.stub = get_unique(block, "stub") === "true" ? true : undefined;
 
     this.item = "table";
   }